refactor(genai-util): split invocations into different files
