+++ packages/genai-util/src/invocations/embedding.ts
@@ -1,0 +1,158 @@
+/*
+ * Copyright The OpenTelemetry Authors
+ * SPDX-License-Identifier: Apache-2.0
+ */
+
+import {
+  SpanStatusCode,
+  type Attributes,
+  type HrTime,
+  type Span,
+} from '@opentelemetry/api';
+import { ATTR_ERROR_TYPE } from '@opentelemetry/semantic-conventions';
+import {
+  ATTR_GEN_AI_OPERATION_NAME,
+  ATTR_GEN_AI_PROVIDER_NAME,
+  ATTR_GEN_AI_REQUEST_MODEL,
+  ATTR_GEN_AI_RESPONSE_MODEL,
+  ATTR_GEN_AI_USAGE_INPUT_TOKENS,
+  GEN_AI_OPERATION_NAME_VALUE_EMBEDDINGS,
+} from '../semconv';
+import type { TokenUsage } from '../types';
+import { calculateDurationSeconds, getErrorType } from '../utils';
+import type { TelemetryHandler } from '../handler';
+
+/**
+ * Manages the lifecycle and telemetry of an Embedding operation.
+ */
+export class EmbeddingInvocation {
+  private readonly _span: Span;
+  private readonly _handler: TelemetryHandler;
+  private readonly _startTime: HrTime;
+  private readonly _providerName: string;
+  private readonly _requestModel?: string;
+  private _responseModel?: string;
+  private _usage?: TokenUsage;
+  private _isEnded = false;
+
+  constructor(
+    span: Span,
+    handler: TelemetryHandler,
+    options: {
+      providerName: string;
+      requestModel?: string;
+      inputTexts?: string[];
+      attributes?: Attributes;
+    },
+    startTime: HrTime = process.hrtime()
+  ) {
+    this._span = span;
+    this._handler = handler;
+    this._startTime = startTime;
+    this._providerName = options.providerName;
+    this._requestModel = options.requestModel;
+
+    const attrs: Attributes = {
+      [ATTR_GEN_AI_PROVIDER_NAME]: this._providerName,
+      [ATTR_GEN_AI_OPERATION_NAME]: GEN_AI_OPERATION_NAME_VALUE_EMBEDDINGS,
+      ...options.attributes,
+    };
+
+    if (this._requestModel) {
+      attrs[ATTR_GEN_AI_REQUEST_MODEL] = this._requestModel;
+    }
+
+    this._span.setAttributes(attrs);
+  }
+
+  public setResponseModel(model: string): this {
+    this._responseModel = model;
+    this._span.setAttribute(ATTR_GEN_AI_RESPONSE_MODEL, model);
+    return this;
+  }
+
+  public getResponseModel(): string | undefined {
+    return this._responseModel;
+  }
+
+  public setUsage(usage: TokenUsage): this {
+    this._usage = usage;
+    if (usage.inputTokens !== undefined) {
+      this._span.setAttribute(
+        ATTR_GEN_AI_USAGE_INPUT_TOKENS,
+        usage.inputTokens
+      );
+    }
+    return this;
+  }
+
+  public setAttribute(key: string, value: any): this {
+    this._span.setAttribute(key, value);
+    return this;
+  }
+
+  public stop(endTime?: HrTime | number): void {
+    if (this._isEnded) {
+      return;
+    }
+    this._isEnded = true;
+    const endHr = Array.isArray(endTime) ? endTime : process.hrtime();
+    const durationSec = calculateDurationSeconds(this._startTime, endHr);
+
+    const metricAttrs: Attributes = {
+      [ATTR_GEN_AI_PROVIDER_NAME]: this._providerName,
+      [ATTR_GEN_AI_OPERATION_NAME]: GEN_AI_OPERATION_NAME_VALUE_EMBEDDINGS,
+    };
+    if (this._requestModel) {
+      metricAttrs[ATTR_GEN_AI_REQUEST_MODEL] = this._requestModel;
+    }
+    if (this._responseModel) {
+      metricAttrs[ATTR_GEN_AI_RESPONSE_MODEL] = this._responseModel;
+    }
+
+    this._handler.recordOperationDuration(durationSec, metricAttrs);
+    if (this._usage) {
+      this._handler.recordTokenUsage(this._usage, metricAttrs);
+    }
+
+    this._span.setStatus({ code: SpanStatusCode.OK });
+    this._span.end(endHr);
+  }
+
+  public fail(
+    error: Error | string | unknown,
+    endTime?: HrTime | number
+  ): void {
+    if (this._isEnded) {
+      return;
+    }
+    this._isEnded = true;
+    const endHr = Array.isArray(endTime) ? endTime : process.hrtime();
+    const durationSec = calculateDurationSeconds(this._startTime, endHr);
+
+    const errorType = getErrorType(error);
+    const metricAttrs: Attributes = {
+      [ATTR_GEN_AI_PROVIDER_NAME]: this._providerName,
+      [ATTR_GEN_AI_OPERATION_NAME]: GEN_AI_OPERATION_NAME_VALUE_EMBEDDINGS,
+      [ATTR_ERROR_TYPE]: errorType,
+    };
+
+    this._handler.recordOperationDuration(durationSec, metricAttrs);
+
+    this._span.setAttribute(ATTR_ERROR_TYPE, errorType);
+
+    if (error instanceof Error) {
+      this._span.recordException(error);
+    }
+
+    this._span.setStatus({
+      code: SpanStatusCode.ERROR,
+      message: error instanceof Error ? error.message : String(error),
+    });
+    this._span.end(endHr);
+  }
+
+  public getSpan(): Span {
+    return this._span;
+  }
+}
